requirements.txt
absl-py==2.1.0
accelerate==0.28.0
anyio==4.3.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
diskcache==5.6.3
exceptiongroup==1.2.0
filelock==3.13.3
flatbuffers==24.3.25
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
inquirerpy==0.3.4
Jinja2==3.1.3
keras==3.1.1
libclang==18.1.1
llama_cpp_python==0.2.58
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
ollama==0.1.8
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pfzy==0.3.4
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
Pygments==2.17.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
safetensors==0.4.2
six==1.16.0
sniffio==1.3.1
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.16.1
termcolor==2.4.0
tokenizers==0.15.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.39.2
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
